feat: 프론트엔드 세팅
