+++ frontend/src/App.tsx
@@ -1,0 +1,9 @@
+function App() {
+  return (
+    <div>
+      <h1>VoteDots</h1>
+    </div>
+  );
+}
+
+export default App;
